backend/requirements.txt
django
djangorestframework
django-filter
django-rest-auth
django-mysql
gunicorn
mysqlclient
